Part 1 of arcan ebuild rewrite
The initial arcan ebuild was written when I had a very poor
understanding of how arcan operated, and what exactly the flags meant.
(see the ffmpeg flag just choosing whether or not to disable the
encode frameserver, as that was a stopgap to fix ffmpeg related
build errors, but it in no means indicates specifically ffmpeg support)

Part 2 of this rewrite will be going through the dependencies on my
minimal arcan machine and correcting them per the use flags, then
writing the proper cmake args for the selected use flags, and that
should support most of the arcan feature set.

If it turns out to be non trivial, a possible part 3 will be
conditionally adding the static libs for LWA, and at this point
I will feel comfortable migrating the work from the live ebuild to
a tagged release.